Created a way to run xml scripts, which for now contain commands for resource preloading (and further commands  which are triggered when loading is finished / or an error has occured) and action dispatching.
Started to implement basic GUI functionality, loading form and xml complete with a few basic attributes.

Started to implement Label, Panel and Button (still needs overlay change for hovering).

+++ bee/src/edu/catalindumitru/bee/concurent/ScheduleProvider.java
@@ -12,21 +12,24 @@ public interface ScheduleProvider {
     /**
      * Schedules a new task. If type is set ti delay, than the task will run once, after the given amount of time. If it
      * set to periodic, then the task will run periodically, once every "amount" time in milliseconds.
+     *
      * @param runnable which task to tun.
-     * @param type type of schedule.
-     * @param amount time in milliseconds of which to repeat / delay the schedule.
+     * @param type     type of schedule.
+     * @param amount   time in milliseconds of which to repeat / delay the schedule.
      */
     public void schedule(Runnable runnable, TYPE type, int amount);
 
     /**
      * Cancels the task from ir's periodic update / delay.
+     *
      * @param runnable which runnable to stop. If no runnable is found, then no action made.
      */
     public void cancel(Runnable runnable);
 
     /**
      * Returns the current time in milliseconds.
+     *
      * @return current time in milliseconds
      */
-    public int currentTime();
+    public double currentTime();
 }
